+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to `songart` will be documented in this file.
 
 ---
 
+## [Unreleased]
+
+### Added
+- [#39](https://github.com/sansoo1972/songart/issues/39) Optional segmented LED-style Spectrum analyzer rendering with dim inactive rows and lit active levels.
+
+---
+
 ## [0.16.0] - 2026-07-18
 
 ### Added